backend_site/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=23.1
psycopg2-binary>=2.9
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.2.2
Pillow>=9.0
django-cors-headers>=4.0
